chore(deps): vite 5 -> 8, vitest 3 -> 4
Vite and vitest ecosystem leap. Vite 8 is current (we were on 5).
Vitest 4 accepts Vite ^6 || ^7 || ^8.

Config migration (the meat of the PER):
- vitest.workspace.mts is removed in vitest 4; projects inline into
  vitest.config.mts's test.projects with extends: true to inherit
  root-level options.
- poolOptions.forks.singleFork -> top-level isolate: false + maxWorkers: 1
  (the pool rework in vitest 4 flattened pool-specific options).
- rootAt(import.meta) -> process.cwd() in the vitest config: vitest 4
  bundles the config through node_modules/.vite-temp/, which rewrites
  import.meta.url. The previous pathname derivation produced an empty
  glob. process.cwd() sidesteps the rewrite.
- env.d.ts: import "vitest/importMeta.js" -> "vitest/importMeta".
  Vitest 4's package.json exports no longer accept the .js extension
  for the types-only entry, which caused ImportMeta.vitest to stop
  merging and broke in-source tests under no-unsafe-* lints.

No runtime semantics changes needed.

Baselines: 188 pass / 0 fail / 24 skip, types clean, 63/63 lint.
