+++ workspace/test-utils/package.json
@@ -29,10 +29,10 @@
     "@starbeam/shared": "workspace:^",
     "@starbeam/universal": "workspace:^",
     "@starbeam/verify": "workspace:^",
-    "@vitest/expect": "^3.2.4",
-    "@vitest/runner": "^3.2.4",
+    "@vitest/expect": "^4.1.4",
+    "@vitest/runner": "^4.1.4",
     "vite-env": "^1.0.0",
-    "vitest": "^3.2.4"
+    "vitest": "^4.1.4"
   },
   "devDependencies": {
     "@starbeam-dev/compile": "workspace:*",
